### 2021-06-19 #3387 @danielskatz
👋 @ph-rast - I (and whedon) don't see a license for this code. JOSS requires software to have an OSI-approved license for it to be reviewed
