### 2021-06-21 #3387 @jbytecode
Hi @ph-rast

Thank you for the submission.

- As @danielskatz stated, please set a valid OSI-approved license (see the list in https://opensource.org/licenses/category)
- Could you please take a look through https://bit.ly/joss-reviewers to suggest some potential reviewers?
