### 2021-06-22 #3387 @ph-rast
Hi @jbytecode ,
I now include a license, GPL3.0.

Suggested reviewers:
- ChristopherLucas
- pedro-teles-fonseca
- Medewitt
