### 2021-06-22 #3387 @jbytecode
@ph-rast Thank you. This list very helpful as the suggested reviewers have R and Econometrics in their preferred language and research interest.
