### 2021-06-22 #3387 @jbytecode
@ptfonseca Hi, would you willing to assist in reviewing this submission for JOSS?
