### 2021-06-22 #3387 @jbytecode
@ChristopherLucas thank you for the response. The process will start after we set the second reviewer. Meanwhile, you can have a look at the [reviewing guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) of JOSS if you are not familiar with the system.
