### 2021-06-29 #3387 @jbytecode
Dear @zedtaha, I am just pinging you to ask if you would be willing to assist in reviewing this submission for JOSS?
