### 2021-07-02 ae697fc
Added VAR means option
