+++ man/dot-print.summary.means.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/print.R
-\name{.print.summary.arma}
-\alias{.print.summary.arma}
-\title{Print helper for ARMA component.}
+\name{.print.summary.means}
+\alias{.print.summary.means}
+\title{Print helper for means component.}
 \usage{
-.print.summary.arma(bmsum)
+.print.summary.means(bmsum)
 }
 \arguments{
 \item{bmsum}{summary.bmgarch object.}
@@ -13,7 +13,7 @@
 Void.
 }
 \description{
-Print helper for ARMA component.
+Print helper for means component.
 }
 \author{
 Stephen R. Martin, Philippe Rast
