### 2021-07-04 #3387 @jbytecode
Hi @mwt, would you be willing to assist in reviewing this submission for JOSS?

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "bmgarch: An R-Package for Bayesian Multivariate GARCH models". You can find more information at the top of this Github issue.
