### 2021-07-04 #3387 @mwt
Hi @jbytecode. Sure, I'm happy to give it a try. I also just started reviewing #3352 just so you know. I should be able to do both unless I am underestimating how much time this takes.

I am a little concerned by:
> Please add following citation when referring to to this package:
>
> Rast, P., Martin, S. R., Liu, S., & Williams, D. R. (in press). A New Frontier for Studying Within-Person Variability: Bayesian Multivariate Generalized Autoregressive Conditional Heteroskedasticity Models. *Psychological Methods*. [https://doi.org/10.1037/met0000357](https://doi.org/10.1037/met0000357)

Because I thought that you're not supposed to publish code in JOSS that is published elsewhere.
