### 2021-07-04 #3387 @jbytecode
Dear @mwt,

It is not problem to have two distinct reviewing jobs at the same time. The deadline is not rigid so we can wait a little while.

The citation you have mentioned is a research paper. It is usual to have a research published somewhere else and submit a paper about its software and repository in JOSS. Thank you very much for your attention and reminder.
