### 2021-07-04 #3452 @whedon
[REVIEW]: bmgarch: An R-Package for Bayesian Multivariate GARCH models
**Submitting author:** @ph-rast (<a href="http://orcid.org/0000-0003-3630-6629">Philippe Rast</a>)
**Repository:** <a href="https://github.com/ph-rast/bmgarch" target ="_blank">https://github.com/ph-rast/bmgarch</a>
**Version:** v1.0.1

- [x] #3509
