### 2021-07-04 #3452 @mwt
Hi! There are other GARCH estimators in R (eg. in `tseries`), but the SoN makes it sound like this is the first in a non-proprietary language. The R package `bayesGARCH` looks particularly similar. I think that the SoN should explain the differences between this and these other R packages.
