+++ doc/paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
  - name: University of California, Davis
    index: 1
- - name: Comscore
+ - name: Comscore, Inc.
    index: 2
 date: June 18 2021
 bibliography: paper.bib
