+++ doc/paper.bib
@@ -102,6 +102,20 @@ pages = {1939--1463},
 year = {2020}
 }
 
+@Manual{rmgarch,
+    title = {rmgarch: Multivariate GARCH models},
+    author = {Ghalanos, Alexios},
+    year = {2019},
+    note = {R package version 1.3-7.},
+  }
+  
+  @Manual{mgarchBEKK,
+    title = {mgarchBEKK: Simulating, Estimating and Diagnosing MGARCH (BEKK and mGJR) Processes},
+    author = {Schmidbauer, Harald and Roesch, Angi and Tunalioglu, Vehbi Sinan},
+    year = {2016},
+    note = {R package version 0.0.2},
+  }
+
 @Misc{rstan,
     title = {{RStan}: the {R} interface to {Stan}},
     author = {{Stan Development Team}},
