### 2021-07-12 42694c4
Updated the statement of need
Clarified the relation of `bmgarch` with respect to other packages currently available on CRAN, in reaction to @mwt 's comments.
