### 2021-07-12 #3452 @ph-rast
> Hi! There are other GARCH estimators in R (eg. in `tseries`), but the SoN makes it sound like this is the first in a non-proprietary language. The R package `bayesGARCH` looks particularly similar. I think that the SoN should explain the differences between this and these other R packages. I think these other ones may be all univariate.

Sorry - this somehow slipped my attention. Was in issues over at ph-rast/bmgarch. Indeed, as you @mwt noted these packages are all univariate and the bayesian package allows only 1,1 order processes. The Statement of Need section now clarifies these things
