+++ README.md
@@ -6,6 +6,7 @@
 ![build](https://github.com/ph-rast/bmgarch/workflows/build/badge.svg)
 ![R-CMD-check](https://github.com/ph-rast/bmgarch/workflows/R-CMD-check/badge.svg)
 [![codecov](https://codecov.io/gh/ph-rast/bmgarch/branch/master/graph/badge.svg)](https://codecov.io/gh/ph-rast/bmgarch)
+[![status](https://joss.theoj.org/papers/fc0db2c4d3bbf1eb18611131f848ed5a/status.svg)](https://joss.theoj.org/papers/fc0db2c4d3bbf1eb18611131f848ed5a)
 <!-- badges: end -->
 
 bmgarch
@@ -31,9 +32,10 @@ Installation
 
 Linux users may need to install `libv8` prior to installing `bmgarch`.
 For example, in Ubuntu, run `sudo apt install libv8-dev` before
-installing the package from CRAN or github. For those who's distro installs
-`libnode-dev` instead of `libv8-dev`, run `install.packages("V8")` in R
-prior to installing `bmgarch` (during installation`rstan` looks explicitly for V8)
+installing the package from CRAN or github. For those who’s distro
+installs `libnode-dev` instead of `libv8-dev`, run
+`install.packages("V8")` in R prior to installing `bmgarch` (during
+installation`rstan` looks explicitly for V8).
 
 The development version can be installed from
 [GitHub](https://github.com/) with:
@@ -100,66 +102,66 @@ intercept-only model for the means.
     #> ---
     #> Iterations:  1000
     #> Chains:  4
-    #> Date:  Mon Sep 21 17:55:01 2020
-    #> Elapsed time (min):  18.1
+    #> Date:  Mon Jul 12 17:37:47 2021
+    #> Elapsed time (min):  14.38
     #> 
     #> ---
     #> Constant correlation, R (diag[C]*R*diag[C]):
     #> 
-    #>         mean  sd  mdn  2.5% 97.5%  n_eff Rhat
-    #> R_Ng-Ps 0.02 0.5 0.03 -0.92  0.93 833.75 1.01
+    #>         mean   sd  mdn  2.5% 97.5%   n_eff Rhat
+    #> R_Ng-Ps 0.01 0.49 0.01 -0.94  0.91 1464.11    1
     #> 
     #> 
     #> Constant variances (diag[C]):
     #> 
     #>        mean   sd  mdn 2.5% 97.5%  n_eff Rhat
-    #> var_Ps 0.59 0.78 0.28 0.01  2.98 320.54 1.02
-    #> var_Ng 1.23 0.40 1.25 0.37  1.98 435.21 1.01
+    #> var_Ps 0.64 0.84 0.29 0.01  3.18 317.46 1.01
+    #> var_Ng 1.21 0.42 1.26 0.18  1.91 293.43 1.01
     #> 
     #> 
     #> MGARCH(1,1) estimates for A:
     #> 
-    #>         mean   sd  mdn  2.5% 97.5%  n_eff Rhat
-    #> A_Ps-Ps 0.34 0.10 0.34  0.15  0.55 845.81 1.01
-    #> A_Ng-Ps 0.06 0.08 0.07 -0.08  0.20 948.74 1.00
[…]
