### 2021-07-15 5d05ea2
added VAR to the testthat loop
