### 2021-07-18 #3452 @mwt
Hi @ph-rast,

I ran the unit tests provided and there were many successful tests. However, I got an error when it reached the `pdBEKKMGARCH` model. I got the same error when trying to run Example 1 from the docs (which uses this model). The example I'm referring to is:

~~~r
library(bmgarch)

data(panas)

# Fit pdBEKK(1, 1) with ARMA(1,1) on the mean structure.
fit <- bmgarch(panas,
               parameterization = "pdBEKK",
               iterations = 1000,
               P = 1, Q = 1,
               distribution = "Student_t",
               meanstructure = "arma")
~~~

My return from this is

~~~
#>
#> CHECKING DATA AND PREPROCESSING FOR MODEL 'pdBEKKMGARCH' NOW.
#>
#> COMPILING MODEL 'pdBEKKMGARCH' NOW.
#>
#> STARTING SAMPLER FOR MODEL 'pdBEKKMGARCH' NOW.
#> Error in unserialize(socklist[[n]]) : error reading from connection
#> Error in serialize(data, node$con, xdr = FALSE) :
#>   error writing to connection
~~~

This is on Windows in R 4.1.0 using the latest CRAN (not development) version. Please let me know if you think I may be missing a step. If this runs for other people, it's probably just my environment. I can retest in a VM.
