### 2021-07-18 ph-rast/bmgarch#13 @mwt
Community guidelines
Hi,

For openjournals/joss-reviews/issues/3452, you need some sort of community guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

These could be in your readme or in some other file.
