+++ .github/workflows/Unit_testing.yml
@@ -6,7 +6,7 @@
 # See https://github.com/r-lib/actions/tree/master/examples#readme for
 # additional example workflows available for the R community.
 
-name: R-CMD-check
+name: Unit-Testing
 
 on:
   push:
