+++ .github/workflows/Unit_testing.yml
@@ -28,28 +28,40 @@ jobs:
 #        - { os: macOS-latest, r: 'release'}
 #        - { os: ubuntu-16.04,   r: '3.6',   vdiffr: false, xref: true,  cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest" }
       #  r-version: [3.6]
+    env:
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      RSPM: ${{ matrix.config.rspm }}
+      # don't treat missing suggested packages as error
+      _R_CHECK_FORCE_SUGGESTS_: false
 
     steps:
-      # - name: Install system dependencies on Linux
-      #   if: runner.os == 'Linux'
-      #   run: |
-      #     sudo add-apt-repository ppa:ubuntu-toolchain-r/test -y
-      #     sudo apt-get update -q
-      #     sudo apt-get -y install libv8-dev libcurl4-openssl-dev
-      
-      - name: Set time zone for windows
-        if: runner.os == 'Windows'
-        run: tzutil /s "Pacific Standard Time"
       - uses: actions/checkout@v2
-      - name: Set up R ${{ matrix.r-version }}
-        uses: r-lib/actions/setup-r@master
+
+      - uses: r-lib/actions/setup-r@master
         with:
-          r-version: ${{ matrix.r-version }}
+          r-version: ${{ matrix.config.r }}
+
+      - uses: r-lib/actions/setup-pandoc@master
+
+      - name: Query dependencies
+        run: |
+          install.packages('remotes')
+          saveRDS(remotes::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
+          writeLines(sprintf("R-%i.%i", getRversion()$major, getRversion()$minor), ".github/R-version")
+        shell: Rscript {0}
+    
       - name: Install dependencies
         run: |
-          install.packages(c("remotes", "devtools"))
+          remotes::install_deps(dependencies = TRUE)
+          remotes::install_cran("devtools")
+        shell: Rscript {0}
+          
+      - name: build package
+        run: |
+          install.packages( "remotes" )
           remotes::install_deps(dependencies = TRUE)
         shell: Rscript {0}
+
       - name: Testing
         run: devtools::test()
         shell: Rscript {0}
