+++ .github/workflows/Unit_testing.yml
@@ -26,7 +26,7 @@ jobs:
         config:
           - { os: windows-latest, r: 'release'}
           - { os: macOS-latest, r: 'release'}
-          - { os: ubuntu-20.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
+       #   - { os: ubuntu-20.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       RSPM: ${{ matrix.config.rspm }}
@@ -81,13 +81,14 @@ jobs:
           
       - name: build package
         run: |
-          install.packages( "remotes" )
+          install.packages( "remotes", "devtools" )
           remotes::install_deps(dependencies = TRUE)
-        shell: Rscript {0}
-
-      - name: Testing
-        run: |
-          install.packages("devtools")
           devtools::test()
         shell: Rscript {0}
+
+      #- name: Testing
+      #  run: |
+      #    install.packages("devtools")
+      #    devtools::test()
+      #  shell: Rscript {0}
 ...
