+++ .github/workflows/Unit_testing.yml
@@ -1,56 +1,0 @@
-# This workflow uses actions that are not certified by GitHub.
-# They are provided by a third-party and are governed by
-# separate terms of service, privacy policy, and support
-# documentation.
-#
-# See https://github.com/r-lib/actions/tree/master/examples#readme for
-# additional example workflows available for the R community.
-
-name: Unit-Testing
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  build:
-    runs-on: ${{ matrix.config.os }}
-    
-    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-
-    strategy:      
-      fail-fast: false
-      matrix:
-        config:
-          - { os: windows-latest, r: 'devel'}
-       #   - { os: macOS-latest, r: 'release'}
-       #   - { os: ubuntu-20.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
-    env:
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-      RSPM: ${{ matrix.config.rspm }}
-      # don't treat missing suggested packages as error
-      _R_CHECK_FORCE_SUGGESTS_: false
-
-    steps:
-      - uses: actions/checkout@v2
-      - name: Set up R ${{ matrix.r-version }}
-        uses: r-lib/actions/setup-r@master
-        with:
-          r-version: ${{ matrix.r-version }}
-      - name: Install dependencies
-        run: |
-          install.packages(c("remotes", "devtools"))
-          remotes::install_deps(dependencies = TRUE)
-        shell: Rscript {0}
-      - name: Test
-        run: | 
-          fit <- bmgarch(panas,
-                 parameterization = "pdBEKK",
-                 iterations = 100,
-                 P = 1, Q = 1,
-                 distribution = "Student_t",
-                 meanstructure = "arma")
-        shell: Rscript {0}
-...
