### 2021-07-20 #3452 @ph-rast
Thanks @mwt -- I installed the package on a fresh Windows VM with a fresh install of R 4.1.0. The model ran without errors. I also checked on macOS and Ubuntu, where it also runs without error.
I am not sure where the error you get might be coming from to be honest, looks like parallelization maybe failed? You could run fewer iterations on just one chain with `chains = 1`?
