### 2021-07-20 c509598
Include Windows and Ubuntu in R-CMD-check
