### 2021-07-20 #3452 @mwt
That's great. I'll try it on a new Windows VM just to make sure, but I imagine it's fine.
