+++ README.Rmd
@@ -237,3 +237,13 @@ We can generate forecasts given the known values of the predictor. Note that the
 fc2x <- forecast(fitx, ahead = 5, xC = stocks.z[101:105, c("nissan", "nissan")])
 fc2x
 ```
+
+## Community Guidelines
+
+1. Contributions and suggestions to the software are always welcome. Please consult our [contribution guidelines](CONTRIBUTING.md) prior to submitting a pull request.
+2. Report issues or problems with the software using github's [issue tracker](https://github.com/ph-rast/bmgarch/issues).
+3. Contributors must adhere to the [Code of Conduct](CODE_OF_CONDUCT.md).
+
+## Acknowledgment
+This work was supported by the National Institute On Aging of the National Institutes of Health under
+Award Number [R01AG050720](https://reporter.nih.gov/project-details/9336761) to PR. The content is solely the responsibility of the authors and does not necessarily represent the official views of the funding agency.
