### 2021-07-26 ph-rast/bmgarch#13 @ph-rast
Hi @mwt, I added a section "[community Guidelines](https://github.com/ph-rast/bmgarch#community-guidelines)" at the bottom of the README. I've made use of the `tidyverse` templates - let me know if this is what you were looking for.
