### 2021-07-27 #3452 @mwt
I still got the same error on a fresh Windows VM. If I set chains = 1, R crashes. If I set chains higher, each serialized chain crashes. That is, I get four crashes when there are four chains, etc. So, I think the error isn't with serialize, but is with one of the other commands run afterwards. Maybe `rstan`?

**I also tested this in Ubuntu, and it worked.** So, we're only talking about a possible issue with one model on one platform. I think it makes sense for me to check off the functionality box.
