### 2021-07-28 #3452 @stephensrmmartin
@mwt Out of curiosity, what CPU are you currently using?
