### 2021-07-28 #3452 @stephensrmmartin
Ok, I am trying to narrow down exactly why that is occurring.

I am thinking that this is a bug up the chain, further upstream in the Mingw64 gcc compiler shipped in MSYS2; CRAN/R use Rtools40, which in turn use MSYS2 and this gcc compiler.

Based on the comment [here](https://discourse.mc-stan.org/t/rstan-on-windows/16673/16), there is a [bug](https://github.com/msys2/MSYS2-packages/issues/1209) about AVX and variable alignment that is not yet [resolved](https://github.com/msys2/MINGW-packages/pull/8317).

I can reproduce the crash on my windows 10 laptop, but it works fine in the WSL2 subsystem, on ubuntu, and on my arch linux machine (and apparently, on osx based on the CI tests). I think this may be therefore a Windows toolchain problem.
