### 2021-07-28 #3452 @mwt
That makes the most sense to me given that the issue seems to affect only Windows.

I use an Intel i5 with 6 cores.
