+++ README.md
@@ -102,66 +102,66 @@ intercept-only model for the means.
     #> ---
     #> Iterations:  1000
     #> Chains:  4
-    #> Date:  Mon Jul 12 17:37:47 2021
-    #> Elapsed time (min):  14.38
+    #> Date:  Mon Jul 26 14:08:44 2021
+    #> Elapsed time (min):  16.4
     #> 
     #> ---
     #> Constant correlation, R (diag[C]*R*diag[C]):
     #> 
-    #>         mean   sd  mdn  2.5% 97.5%   n_eff Rhat
-    #> R_Ng-Ps 0.01 0.49 0.01 -0.94  0.91 1464.11    1
+    #>         mean   sd  mdn  2.5% 97.5% n_eff Rhat
+    #> R_Ng-Ps 0.15 0.46 0.31 -0.84  0.89 28.91 1.07
     #> 
     #> 
     #> Constant variances (diag[C]):
     #> 
     #>        mean   sd  mdn 2.5% 97.5%  n_eff Rhat
-    #> var_Ps 0.64 0.84 0.29 0.01  3.18 317.46 1.01
-    #> var_Ng 1.21 0.42 1.26 0.18  1.91 293.43 1.01
+    #> var_Ps 0.56 0.73 0.31 0.02  3.04 214.26 1.04
+    #> var_Ng 1.39 0.38 1.48 0.46  1.94  11.14 1.15
     #> 
     #> 
     #> MGARCH(1,1) estimates for A:
     #> 
-    #>         mean   sd  mdn  2.5% 97.5%   n_eff Rhat
-    #> A_Ps-Ps 0.34 0.10 0.34  0.13  0.55  797.05    1
-    #> A_Ng-Ps 0.06 0.08 0.06 -0.09  0.21 1884.70    1
-    #> A_Ps-Ng 0.05 0.14 0.05 -0.25  0.32 2100.56    1
-    #> A_Ng-Ng 0.40 0.12 0.40  0.16  0.63  967.36    1
+    #>         mean   sd  mdn  2.5% 97.5%  n_eff Rhat
+    #> A_Ps-Ps 0.34 0.09 0.33  0.15  0.53 502.52 1.01
+    #> A_Ng-Ps 0.06 0.07 0.06 -0.08  0.20 622.30 1.02
+    #> A_Ps-Ng 0.09 0.14 0.10 -0.22  0.31   6.68 1.24
+    #> A_Ng-Ng 0.37 0.12 0.37  0.06  0.61 222.32 1.03
     #> 
     #> 
     #> MGARCH(1,1) estimates for B:
     #> 
-    #>          mean   sd   mdn  2.5% 97.5%  n_eff Rhat
-    #> B_Ps-Ps  0.75 0.22  0.84  0.12  0.95 300.58 1.01
-    #> B_Ng-Ps -0.10 0.15 -0.09 -0.50  0.19 225.38 1.02
-    #> B_Ps-Ng  0.29 0.36  0.28 -0.57  1.08 805.29 1.01
-    #> B_Ng-Ng  0.36 0.20  0.35  0.03  0.77 705.36 1.00
+    #>          mean   sd   mdn  2.5% 97.5% n_eff Rhat
+    #> B_Ps-Ps  0.77 0.21  0.87  0.12  0.94 55.70 1.07
+    #> B_Ng-Ps -0.10 0.13 -0.13 -0.39  0.19 55.40 1.05
+    #> B_Ps-Ng  0.20 0.38  0.19 -0.67  1.05 20.46 1.09
+    #> B_Ng-Ng  0.31 0.18  0.31  0.02  0.69 84.81 1.07
     #> 
     #> 
[…]
