### 2021-08-03 0bc4565
downgraded requirement of some packages
