+++ DESCRIPTION
@@ -17,8 +17,8 @@ Description: Fit Bayesian multivariate GARCH models using 'Stan' for full Bayesi
 License: GPL (>=3)
 Depends: 
     methods,
-    R (>= 4.1.0),
-    Rcpp (>= 1.0.6)
+    R (>= 4.0.2),
+    Rcpp (>= 1.0.5)
 Imports: 
     rstan (>= 2.21.2),
     rstantools (>= 2.1.1),
@@ -29,8 +29,8 @@ Imports:
     Rdpack
 LinkingTo: 
     BH (>= 1.75.0-0),
-    Rcpp (>= 1.0.6),
-    RcppEigen (>= 0.3.3.9.1),
+    Rcpp (>= 1.0.5),
+    RcppEigen (>= 0.3.3.7.0),
     rstan (>= 2.21.2),
     StanHeaders (>= 2.21.0)
 RdMacros: Rdpack
@@ -40,5 +40,5 @@ NeedsCompilation: yes
 SystemRequirements: GNU make
 RoxygenNote: 7.1.1
 Suggests: 
-    testthat (>= 3.0.2)
+    testthat (>= 2.3.2)
 BugReports: https://github.com/ph-rast/bmgarch/issues
