+++ DESCRIPTION
@@ -28,7 +28,7 @@ Imports:
     loo,
     Rdpack
 LinkingTo: 
-    BH (>= 1.75.0-0),
+    BH (>= 1.72.0-0),
     Rcpp (>= 1.0.5),
     RcppEigen (>= 0.3.3.7.0),
     rstan (>= 2.21.2),
