### 2021-08-04 #3452 @ChristopherLucas
I finished reviewing the package. I had no problem installing the package (Ubuntu 21.04 and R 4.1.0) and all the tests ran without error. You might want to add the docs as a readme so people installing from CRAN find them faster, but if you have reasons to not do so, that seems fine too.

Cool project!
