### 2021-08-05 #3452 @jbytecode
Dear @ph-rast

In lines 43,44, and 45 of paper, Stan Development Team (2016) and Stan Development Team (2020) should be cited as (Stan Development Team, 2016) and (Stan Development Team, 2020) respectively. Please correct the paper.

Finally, please re-check the whole paper by means of both language and format. When it is finished, let whedon generate the pdf and check references here. After all please

- [ ] make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.

Thank you in advance.
