+++ doc/paper.md
@@ -41,11 +41,11 @@ The package is designed to take multivariate time series only. Hence, in terms o
 The default behavior in `bmgarch` is to estimate a CCC(1,1) parameterized model of order 1 for the ARCH and GARCH components assuming Student-$t$ distributed innovations and a constant means (zero, typically) structure. The model is fit using 4 chains with 2000 iterations each (half of those are warm-up and the other half are sampling).
 As such, the minimum call will be `bmgarch( data = X )` with `X` being a matrix or data frame with at least 2 columns.
 
-In order to include the effect of an external predictor variable on the unconditional and constant covariance term $C$, one can include a predictor `Y` of the same dimension as `X` with `bmgarch( data = X, xC = Y )`. Further, `bmgarch` takes a number of model related arguments governing the type of parameterization (`parameterization = {"CCC", "DCC", "BEKK", "pdBEKK"}`), the order of the GARCH (`P = {1, ..., t-1}`) and ARCH (`Q = {1, ..., t-1}`) processes, the mean structure (`meanstructure = {"constant", "arma"}`) and the type of distribution (`distribution = {"Student_t", "Gaussian"}`). The function also takes arguments with respect to sampling (number of `iterations` and number of `chains`) as well as whether data should be z-standardized before sampling (`standardize_data = {FALSE, TRUE}`) -- this last argument can facilitate the estimation process. 
+In order to include the effect of an external predictor variable on the unconditional and constant covariance term $C$, one can include a time-varying predictor `Y` of the same dimension as `X` by including the preditor with `xC = Y`, resulting in`bmgarch( data = X, xC = Y )`. Further, `bmgarch` takes a number of model related arguments governing the type of parameterization (`parameterization = {"CCC", "DCC", "BEKK", "pdBEKK"}`), the order of the GARCH (`P = {1, ..., t-1}`) and ARCH (`Q = {1, ..., t-1}`) processes, the mean structure (`meanstructure = {"constant", "arma"}`) and the type of distribution (`distribution = {"Student_t", "Gaussian"}`). The function also takes arguments with respect to sampling (number of `iterations` and number of `chains`) as well as whether data should be z-standardized before sampling (`standardize_data = {FALSE, TRUE}`) -- this last argument can facilitate the estimation process. 
 
 Objects of the `bmgarch` family can be passed to the `plot()` function, `print()`, and `summary()` for a summary table.
 
-Moreover, each estimated model can be passed to the `forecasting()` function to generate $m$-ahead forecasts of variances, correlations, and, if the ARMA(1,1) is enabled, means. 
+Moreover, each estimated model can be passed to the `forecasting()` function to generate $m$-ahead forecasts of variances, correlations, and, if the meanstructure is included (such as with ARMA(1,1)), forecasts of means. 
 Moreover, `bmgarch` integrates model ensemble techniques to generate model weighted forecasts (and estimates) based on Bayesian model averaging and stacking techniques described by @Yao2018 and @Buerkner2020. The forecasted objects can again be plotted and summarized via the corresponding generic functions. 
 
 # Summary
